+++ chainer_compiler/elichika/typing/ext/chainer_functions.py
@@ -8,10 +8,10 @@ import math
 from   chainer.utils.conv import get_conv_outsize
 from   chainer.utils import type_check
 
-from   chainer_compiler.elichika.typing.ext_functions_utils import *
-from   chainer_compiler.elichika.typing.shape_elem          import *
-from   chainer_compiler.elichika.typing.types               import *
-from   chainer_compiler.elichika.typing.utils               import all_same
+from   chainer_compiler.elichika.typing.ext.utils  import *
+from   chainer_compiler.elichika.typing.shape_elem import *
+from   chainer_compiler.elichika.typing.types      import *
+from   chainer_compiler.elichika.typing.utils      import all_same
 
 
 class ty_ChainerVariable():
