+++ chainer_compiler/elichika/typing/ext/pytorch/utils.py
