+++ chainer_compiler/elichika/typing/ext/pytorch_functions.py
@@ -4,154 +4,12 @@ import torch.nn.functional as F
 
 from   chainer.utils import type_check
 
-from   chainer_compiler.elichika.typing.ext_functions_utils import *
-from   chainer_compiler.elichika.typing.types               import *
-from   chainer_compiler.elichika.typing.pytorch.nn          import *
-from   chainer_compiler.elichika.typing.pytorch.tensor      import *
+from   chainer_compiler.elichika.typing.ext.utils          import *
+from   chainer_compiler.elichika.typing.types              import *
+from   chainer_compiler.elichika.typing.ext.pytorch.nn     import *
+from   chainer_compiler.elichika.typing.ext.pytorch.tensor import *
 
 __all__ = [ 'pytorch_func_ty', 'pytorch_callable_ty' ]
-
-
-class ty_ChainerSum():
-    def __call__(self, ty_args, ty_kwargs):
-        x_type, = ty_args
-        self.dim, lacks_dim = get_kwarg(ty_kwargs, 'dim', default=None)
-        self.keepdims, lacks_keepdims = \
-                get_kwarg(ty_kwargs, 'keepdims', default=False)
-
-        if isinstance(self.dim, int):
-            self.dim = (self.dim,)
-
-        self.check_type_forward(make_multiple_tc_variable(ty_args, ('x',)))
-
-        if self.dim is None:
-            self.dim = tuple(range(x_type.ndim))
-
-        return self.infer_return(x_type)
-
-    def check_type_forward(self, in_types):
-        type_check.expect(in_types[0].dtype.kind == 'f')
-
-        if self.dim is None:
-            return
-
-        for dim in self.dim:
-            if dim >= 0:
-                type_check.expect(
-                    dim < in_types[0].ndim,
-                )
-            else:
-                type_check.expect(
-                    -dim - 1 < in_types[0].ndim,
-                )
-
-    def infer_return(self, x_type):
-        if self.keepdims:
-            ret_shape = list(x_type.shape)
-            for i in self.dim:
-                ret_shape[i] = 1
-            return TyChainerVariable(x_type.dtype, shape=ret_shape)
-
-        ret_shape = remove_dims(x_type.shape, self.dim)
-        return TyChainerVariable(x_type.dtype, shape=ret_shape)
-
-
-
-class ty_ChainerPadSequence():
-    def __call__(self, ty_args, ty_kwargs):
-        xs_type, = ty_args
-        self.length, lacks_length = get_kwarg(ty_kwargs, 'length', None)
-
-        if not xs_type.is_fixed_len:
-            ret_shape = list((None,) * (xs_type.get().ndim + 1))
-            if not lacks_length:
-                ret_shape[1] = self.length
-            return TyChainerVariable(xs_type.get().dtype, shape=ret_shape)
-
-        self.check_type_forward(type_check.make_variable(xs_type, 'xs'))
-        return self.infer_return(xs_type, lacks_length)
-
-    def check_type_forward(self, xs_type):
-        for i in range(xs_type.size().eval()):
-            type_check.expect(
-                xs_type[i].ndim > 0,
-                xs_type[i].shape[1:] == xs_type[0].shape[1:],
-                xs_type[i].dtype == xs_type[0].dtype
-            )
-
-    def infer_return(self, xs_type, lacks_length):
-        n = ShapeElem(xs_type.size())
-        ret_shape = list((n,) + xs_type.get().shape)
-
-        if lacks_length:
-            ret_shape[1] = None
-            return TyChainerVariable(xs_type.get().dtype, shape=ret_shape)
-        if self.length is not None:
-            ret_shape[1] = self.length
-            return TyChainerVariable(xs_type.get().dtype, shape=ret_shape)
-
-        shape_0s = [t.shape[0] for t in xs_type.get_tys()]
-        ret_shape[1] = max(shape_0s)
-        return TyChainerVariable(xs_type.get().dtype, shape=ret_shape)
-
-
-class ty_ChainerEmbedID():
-    def __call__(self, embed, ty_args, ty_kwargs):
-        assert isinstance(ty_args[0], TyTensor)
-        x_type, = ty_args
-
-        assert x_type.dtype.kind == 'i'
-        assert x_type.ndim >= 1
-        ret_shape = x_type.shape + (ShapeElem(embed.W.shape[1]),)
-
-        if not is_incomplete_shape(x_type.shape):
-            assert all([t < embed.W.shape[0] for t in x_type.shape])
-        return TyTorchTensor(embed.W.dtype, shape=ret_shape)
-
-
-class ty_ChainerNStepBiLSTM():
-    def __call__(self, nblstm, ty_args, ty_kwargs):
-        hx_type, cx_type, xs_type = ty_args
-        assert isinstance(xs_type, TySequence)
-        xs_len = ShapeElem(xs_type.size())
-
-        if isinstance(hx_type, TyTensor):
-            hx_shape = hx_type.shape
-            hx_dtype = hx_type.dtype
-        else:
-            hx_shape = (nblstm.n_layers * 2, xs_len, nblstm.out_size)
-            hx_dtype = xs_type.get().dtype
-
-        if isinstance(cx_type, TyTensor):
-            cx_shape = cx_type.shape
-            cx_dtype = cx_type.dtype
-        else:
-            cx_shape = (nblstm.n_layers * 2, xs_len, nblstm.out_size)
-            cx_dtype = hx_dtype
-
-        hy_type = TyChainerVariable(hx_dtype, shape=hx_shape)
-        cy_type = TyChainerVariable(cx_dtype, shape=cx_shape)
-
-        assert hx_shape[0] // 2 == nblstm.n_layers
-        assert hx_shape == cx_shape
-        N = hx_shape[2]
-
-        if not xs_type.is_fixed_len:
-            # TODO
-            ys_shape = (xs_type.get().shape[0], 2 * N)
-            ys_type = TyList(TyChainerVariable(xs_type.get().dtype, shape=ys_shape))
-            return TyTuple([hy_type, cy_type, ys_type])
-
-        xs_dtypes = [t.dtype for t in xs_type.get_tys()]
-        xs_shapes = [t.shape for t in xs_type.get_tys()]
-        assert all_same(xs_dtypes)
-
-        # TODO(momohatt): nblstm doesn't have attribute in_size
-        # assert all([i == nblstm.in_size for _, i in xs_shapes])
-        ys_shapes = [(l, 2 * N) for l, _ in xs_shapes]
-        ys_type = TyList([TyChainerVariable(xs_dtypes[0], shape=s) for s in ys_shapes])
-
-        return TyTuple([hy_type, cy_type, ys_type])
 
 
 pytorch_func_ty = {
