+++ chainer_compiler/elichika/typing/ext/utils.py
